% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digest.R
\name{cor_matrix_samples}
\alias{cor_matrix_samples}
\title{cor_matrix_samples
Create a plot showing the correlation between RNA and Protein in the individual samples.}
\usage{
cor_matrix_samples(gathered_data)
}
\arguments{
\item{gathered_data}{A single tibble containing both protein and data for future processing}
}
\value{
A plot showing the correlation between RNA and Protein levels in individual samples
}
\description{
cor_matrix_samples
Create a plot showing the correlation between RNA and Protein in the individual samples.
}
\examples{
cor_matrix_samples(gathered_data)
}
