% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digest.R
\name{correlation_plot_levels}
\alias{correlation_plot_levels}
\title{correlation_plot_levels
Create a plot showing the correlation between RNA and Protein in each gene.}
\usage{
correlation_plot_levels(correlations_df)
}
\arguments{
\item{correlations_df}{A dataframe of correlations between RNA and Protein levels.}
}
\value{
A plot showing the correlation between RNA and Protein levels in all genes and a dataframe of the 10 most positively and negatively correlated RNA/Protein pairs
}
\description{
correlation_plot_levels
Create a plot showing the correlation between RNA and Protein in each gene.
}
