% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digest.R
\name{create_pw_df}
\alias{create_pw_df}
\title{create_pw_df
Creates a dataframe of the average correlation between RNA and Protein levels within all KEGG pathways detected in the dataset.}
\usage{
create_pw_df(corr_df)
}
\arguments{
\item{corr_df}{A dataframe of correlations between RNA and Protein levels.}
}
\value{
A dataframe containing pathways and the average correlation found in them
}
\description{
create_pw_df
Creates a dataframe of the average correlation between RNA and Protein levels within all KEGG pathways detected in the dataset.
}
