% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table.R
\name{create_table}
\alias{create_table}
\title{Reverse engineer a table from its description from a SQL query}
\usage{
create_table(tbl_name)
}
\arguments{
\item{tbl_name}{Table with holds the value with both the TABLE_NAME as
well as the COLUMN_NAME. See details and examples}
}
\value{
a tibble assigned to the GlobalEnv
}
\description{
Reverse engineer a table from its description from a SQL query
}
\details{
The table to create derives its name from the TABLE_NAME column and the columns from the
COLUMN_NAME column. Overall, the output comes from an Oracle DB
}
\examples{
# mockup table descritpion
(tbl_descr <- tibble::tibble(TABLE_NAME = c("my_table"),
                             COLUMN_NAME = c("var1", "var2", "var3")))
# now print the table
\dontrun{create_table(tbl_descr)}
}
