% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dict_hierarchie}
\alias{dict_hierarchie}
\title{dict_hierarchie}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 27 rows and 3 columns.
}
\usage{
dict_hierarchie
}
\description{
dict_tbl about pep_hierarchie dataset
}
\keyword{datasets}
