% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dict_pepzeiten}
\alias{dict_pepzeiten}
\title{dict_pepzeiten}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 32 rows and 6 columns.
}
\usage{
dict_pepzeiten
}
\description{
dict_tbl about pep_zeiten dataset
}
\keyword{datasets}
