% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_wrap_starfish.R
\name{facet_custom}
\alias{facet_custom}
\alias{facet_wrap_starfish}
\alias{facet_wrap_sf}
\title{STARFISH customized facets with units on right side}
\usage{
facet_wrap_starfish(
  facets = ~anon_unit,
  ncol = 1,
  nrow = NULL,
  strip.position = "right",
  drop = TRUE,
  scales = "fixed",
  shrink = TRUE,
  labeller = "label_value",
  as.table = TRUE,
  switch = NULL,
  dir = "h",
  ...
)

facet_wrap_sf(
  facets = ~anon_unit,
  ncol = 1,
  nrow = NULL,
  strip.position = "right",
  drop = TRUE,
  scales = "fixed",
  shrink = TRUE,
  labeller = "label_value",
  as.table = TRUE,
  switch = NULL,
  dir = "h",
  ...
)
}
\arguments{
\item{facets}{defaults to variable anon_unit}

\item{ncol}{defaults to 1}

\item{nrow}{defaults to NULL}

\item{strip.position}{"right"}

\item{drop}{default: TRUE}

\item{scales}{default: fixed}

\item{shrink}{default fixed}

\item{labeller}{"label_value"}

\item{as.table}{TRUE}

\item{switch}{NULL}

\item{dir}{"h"}

\item{...}{to pass to \code{ggplot2::facet_wrap}}
}
\value{
a facet
}
\description{
STARFISH customized facets with units on right side
}
