% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_shifts_where.R
\name{filter_shifts_where}
\alias{filter_shifts_where}
\title{Filter all shifts where a certain condition is met}
\usage{
filter_shifts_where(.data, ..., shift_var = shift_id)
}
\arguments{
\item{.data}{a data frame or tibble}

\item{...}{condition to pass to dplyr::filter}

\item{shift_var}{variable with the shift_identified, default = "shift_id"}
}
\value{
a dataframe containing all shift-entries where a condition is met
}
\description{
Filter all shifts where a certain condition is met
}
