% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digest.R
\name{gather_rna_prot_data}
\alias{gather_rna_prot_data}
\title{gather_rna_prot_data(rna, protein)
Transforms the data to fit the following functions}
\usage{
gather_rna_prot_data(rna, protein)
}
\arguments{
\item{rna}{a tibble containing mRNA expression values for genes across patient samples}

\item{protein}{a tibble containing protein expression values for genes across patient samples}
}
\value{
a single tibble containing both protein and data for future processing
}
\description{
gather_rna_prot_data(rna, protein)
Transforms the data to fit the following functions
}
