% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_date_starfish.R
\name{geom_covid_starfish}
\alias{geom_covid_starfish}
\title{Highlighting Covid timeframe in NH}
\usage{
geom_covid_starfish(
  date1 = "2020-03-17",
  date2 = "2021-12-01",
  color = "grey30",
  linetype = "dashed",
  alpha = 0.6
)
}
\arguments{
\item{date1}{default set}

\item{date2}{default set}

\item{color}{default set}

\item{linetype}{default set}

\item{alpha}{default set}
}
\value{
a custom geom
}
\description{
Highlighting Covid timeframe in NH
}
