% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digest.R
\name{get_all_metadata_correlations}
\alias{get_all_metadata_correlations}
\title{get_all_metadata_correlations
Produces all plots for RNA and Protein levels related to all metadata variables.}
\usage{
get_all_metadata_correlations(
  metadata,
  gathered_data,
  gene_name,
  regression = T
)
}
\arguments{
\item{metadata}{A tibble containing metadata for each patient in the sample.}

\item{gathered_data}{A single tibble containing both protein and data for processing}

\item{gene_name}{The name of a single gene to plot}

\item{regression}{FALSE/TRUE whether a linear regression should be plotted for continuous metadata parameters}
}
\value{
Returns a list of plots of all the metadata.
}
\description{
get_all_metadata_correlations
Produces all plots for RNA and Protein levels related to all metadata variables.
}
