% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab.R
\name{grab}
\alias{grab}
\title{Grab studies from cbioportal with the specifcied type of data}
\usage{
grab()
}
\arguments{
\item{n}{Minimum number of samples required}

\item{...}{}
}
\value{
Table of candidate studies
}
\description{
Grab studies from cbioportal with the specifcied type of data
}
\examples{
grab()
}
