% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digest.R
\name{KEGG_correlation_plot}
\alias{KEGG_correlation_plot}
\title{KEGG_correlation_plot
Create a plot showing the correlation between RNA and Protein in KEGG pathways.}
\usage{
KEGG_correlation_plot(combined_pw_data)
}
\arguments{
\item{combined_pw_data}{A dataframe containing pathways and the average correlation found in them}
}
\value{
a plot of the average correlation between RNA and Protein levels in all pathways in detected in the gene set
}
\description{
KEGG_correlation_plot
Create a plot showing the correlation between RNA and Protein in KEGG pathways.
}
