% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digest.R
\name{metadata_correlations}
\alias{metadata_correlations}
\title{metadata_correlations}
\usage{
metadata_correlations(
  metadata,
  gathered_data,
  clin_var,
  gene_name,
  regression = T
)
}
\arguments{
\item{metadata}{A tibble containing metadata for each patient in the sample.}

\item{gathered_data}{A single tibble containing both protein and data for processing}

\item{clin_var}{The name of a clinical variable present in the metadata file.}

\item{gene_name}{The name of a single gene to plot}

\item{regression}{FALSE/TRUE whether a linear regression should be plotted for continuous metadata parameters}
}
\value{
A plot showing the protein and RNA levels in relation to a given
}
\description{
Plots the RNA and protein levels in relation to a given clinical parameter. Both gene name and metadata parameter name must be supplied.
}
\details{
The function automatically determines the plot to use; A Conditional Density plot for binary values, A barplot with beeswarm points for non-binary categorical values and finally a regular point plot with or without linear regression for continuous data.
}
