% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digest.R
\name{plot_gene}
\alias{plot_gene}
\title{plot_gene
Plots the RNA/Protein levels for a given gene.}
\usage{
plot_gene(gathered_data, gene_name)
}
\arguments{
\item{gathered_data}{A single tibble containing both protein and data for processing}

\item{gene_name}{The name of a single gene to plot}
}
\value{
A plot showing RNA/Protein levels for the gene provided
}
\description{
plot_gene
Plots the RNA/Protein levels for a given gene.
}
\examples{
plot_gene(gathered_data, "ARPC1A")
}
