% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release_data.R
\name{release_data}
\alias{release_data}
\title{Extract the molecular data from a study}
\usage{
release_data(study_id, molecular_type)
}
\arguments{
\item{study_id}{The cbioportal study ID}

\item{molecular_type}{The type of data to extract}
}
\value{
Tibble with the selected molecular data
}
\description{
Extract the molecular data from a study
}
\examples{
release_data(study_id = "coad_silu_2022", molecular_type = "mrnaRnaSeqSampleCount")
release_data(study_id = "brca_cptac_2020", molecular_type = "mrnaRnaSeqV2SampleCount")
release_data(study_id = "brca_cptac_2020", molecular_type = "massSpectrometrySampleCount")
}
