% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release_metadata.R
\name{release_metadata}
\alias{release_metadata}
\title{Extract the metadata from a study}
\usage{
release_metadata(study_id)
}
\arguments{
\item{study_id}{The cbioportal study ID}
}
\value{
Tibble with the study metadata
}
\description{
Extract the metadata from a study
}
