% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra-palette.R
\name{scale_custom}
\alias{scale_custom}
\alias{scale_color_discrete_starfish}
\alias{scale_colour_discrete_starfish}
\alias{scale_fill_discrete_starfish}
\alias{scale_color_opinionated_starfish}
\alias{scale_colour_opinionated_starfish}
\alias{scale_fill_opinionated_starfish}
\alias{scale_color_diverging_starfish}
\alias{scale_colour_diverging_starfish}
\alias{scale_fill_diverging_starfish}
\alias{scale_color_continuous_starfish}
\alias{scale_colour_continuous_starfish}
\alias{scale_fill_continuous_starfish}
\title{Custom color and fill scales}
\usage{
scale_color_discrete_starfish(palette = "starfish", extend = FALSE, ...)

scale_colour_discrete_starfish(palette = "starfish", extend = FALSE, ...)

scale_fill_discrete_starfish(palette = "starfish", extend = FALSE, ...)

scale_color_opinionated_starfish(palette = "starfish", ...)

scale_colour_opinionated_starfish(palette = "starfish", extend = FALSE, ...)

scale_fill_opinionated_starfish(palette = "starfish", ...)

scale_color_diverging_starfish(palette = "starfish", ...)

scale_colour_diverging_starfish(palette = "starfish", ...)

scale_fill_diverging_starfish(palette = "starfish", ...)

scale_color_continuous_starfish(palette = "starfish", ...)

scale_colour_continuous_starfish(palette = "starfish", ...)

scale_fill_continuous_starfish(palette = "starfish", ...)
}
\arguments{
\item{palette}{Name of color palette}

\item{extend}{Whether to extend discrete color palette to make sufficient colors for levels needed}

\item{...}{Additional arguments to be passed to internal scale function}
}
\description{
Custom coloring and filling functions based on unique color palettes
}
\details{
Specific functions include:
\itemize{
\item{scale_(color/colour/fill)\emph{discrete_starfish}{Discrete palette with either fixed or dynamically extended number of shades}
\item{scale}(color/colour/fill)\emph{opinionated_starfish}{Discrete palette with specific values for "good", "bad", and "neutral"}
\item{scale}(color/colour/fill)\emph{diverging_starfish}{Continuous diverging color palette}
\item{scale}(color/colour/fill)_continuous}{COntinuous color palette}
}
}
