% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\docType{data}
\name{sf_colors}
\alias{sf_colors}
\title{A collection of colors}
\format{
An object of class \code{character} of length 5.
}
\usage{
sf_colors
}
\description{
what about fonts??
}
\keyword{datasets}
