% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squeeze.R
\name{squeeze}
\alias{squeeze}
\title{Extract studies from cbioportal with the specifcied type of data and minimum number of samples}
\usage{
squeeze(n, ...)
}
\arguments{
\item{n}{Minimum number of samples}

\item{...}{Data types}
}
\value{
Tibble with all studies
}
\description{
Extract studies from cbioportal with the specifcied type of data and minimum number of samples
}
