% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starfish_all.r
\name{starfish_all}
\alias{starfish_all}
\title{starfish_all}
\usage{
starfish_all(
  sv_file,
  cnv_file,
  gender_file,
  prefix_total = "",
  genome_version = "hg19",
  plot = TRUE,
  smethod = "class"
)
}
\arguments{
\item{sv_file}{a SV dataframe with 8 columns: "chrom1","end1", "chrom2","end2","svtype" (DEL,DUP,h2hINV,t2tINV,TRA),"strand1" (+/-) and "strand2" (+/-),"sample". Other svtypes like INV, INS, BND are not accepted}

\item{cnv_file}{a CNV dataframe with 5 columns: "chromosome","start","end","total_cn", and "sample". "total_cn" should contain absolute copy numbers.}

\item{gender_file}{a sample table with 2 columns: "sample" and "gender" ("Female, "female","F","f","Male","male","M","m"). If gender is unknown, any other characters could be put here like "unknown".}

\item{prefix_total}{the prefix for all intermediate files, default is none}

\item{genome_version}{which genome assembly was used to call SV and CNV. It should be "hg19" or "hg38", default is "hg19"}

\item{plot}{the logical value of plotting "connected" CGRs, default is TRUE}

\item{smethod}{"class" based on a pre-constructed classifier from PCAWG dataset or "cluster" based on de-novo unsupervised clustering, default is "class"}
}
\value{
a list of files: $complex_sv contains complex SVs, $starfish_call contains connected CGRs, $feature_matrix contains CGR feature matrix, and $chrss_signature shows the signatures computed and inferred from PCAWG dataset. The signature classification table and plot will be generated if "class" is selected, otherwise clustering matrices and plots will be stored under "CGR_cluster" folder with K from 2 to 10.
}
\description{
This function loads SV, CNV and gender files, identifies "connected" complex genome rearrangement (CGR) regions and complex SVs, then constructs a CGR event vs. feature matrix and infers CGR signatures based on clustering and classification.
}
