% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starfish_feature.r
\name{starfish_feature}
\alias{starfish_feature}
\title{starfish_feature}
\usage{
starfish_feature(
  cgr,
  complex_sv,
  cnv_file,
  gender_file,
  prefix = "",
  genome_v = "hg19",
  cnv_factor = "auto",
  arm_del_rm = TRUE
)
}
\arguments{
\item{cgr}{"connected" CGR regions, which is the output of starfish_link_out$starfish_call}

\item{complex_sv}{complex SVs, which is the output from starfish_link_out$interleave_tra_complex_sv}

\item{cnv_file}{a CNV dataframe with 5 columns: "chromosome","start","end","total_cn", and "sample". "total_cn" should contain absolute copy numbers}

\item{gender_file}{a sample table with 2 columns: "sample" and "gender". Gender could be "Female, "female","F","f","Male","male","M", or "m". If the gender is unknown, any other characters could be given, such as "unknown", and the gender will be inferred by the CN baseline of chromosome X}

\item{prefix}{the prefix for all intermediate files, default is none}

\item{genome_v}{which genome assembly was used to call SV and CNV. It should be "hg19" or "hg38", default is "hg19"}

\item{cnv_factor}{the CN fluctuation beyond or below baseline to identify loss and gain fragments for samples with decimal CN, default is "auto", or users can provide a value between 0 and 1}

\item{arm_del_rm}{the logical value of removing arm level deletion or not, default is TRUE}
}
\value{
a list of files: $cluster_feature is the CGR region vs. feature matrix,$cnv_baseline is the cnv file with baseline annotation
}
\description{
This function loads "connected" CGR regions and complex SVs reported by starfish_link, then combines CNV calls and sample gender to construct a CGR region vs. feature matrix for downstream clustering and classification.
}
