% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starfish_link.r
\name{starfish_link}
\alias{starfish_link}
\title{starfish_link}
\usage{
starfish_link(sv_file, prefix = "")
}
\arguments{
\item{sv_file}{a SV dataframe with 8 columns: "chrom1","pos1", "chrom2","pos2","svtype" (DEL,DUP,h2hINV,t2tINV,TRA),"strand1" (+/-) and "strand2" (+/-),"sample". Other svtypes like INV, INS, BND are not accepted}

\item{prefix}{the prefix for all intermediate files, default is none}
}
\value{
a list of files: $interleave_tra_complex_sv contains complex SVs, $mergecall contains "seed" CGR regions, and $starfish_call contains "connected" CGR regions.
}
\description{
This function loads a SV dataframe and generates "seed" CGR regions, "connected" CGR regions, and complex SVs.
}
