% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\docType{data}
\name{starfish_pal}
\alias{starfish_pal}
\alias{starfish_pal_op}
\alias{starfish_pal_cont}
\alias{starfish_pal_div}
\title{STARFISH palette
with six dimensions}
\format{
An object of class \code{character} of length 17.

An object of class \code{character} of length 3.

An object of class \code{character} of length 2.

An object of class \code{character} of length 3.
}
\usage{
starfish_pal

starfish_pal_op

starfish_pal_cont

starfish_pal_div
}
\description{
STARFISH palette
with six dimensions
}
\keyword{datasets}
