% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\docType{data}
\name{starfish_pal2}
\alias{starfish_pal2}
\title{Pastel palette}
\format{
An object of class \code{character} of length 5.
}
\usage{
starfish_pal2
}
\description{
Pastel palette
}
\keyword{datasets}
