% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\docType{data}
\name{starfish_pal_ten}
\alias{starfish_pal_ten}
\title{Ten-color palette}
\format{
An object of class \code{character} of length 10.
}
\usage{
starfish_pal_ten
}
\description{
Ten-color palette
}
\keyword{internal}
