% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starfish_plot.r
\name{starfish_plot}
\alias{starfish_plot}
\title{starfish_plot}
\usage{
starfish_plot(sv_file, cnv_file, cgr, genome_v = "hg19")
}
\arguments{
\item{sv_file}{a SV dataframe with "chrom1","end1", "chrom2","end2","svtype","strand1","strand2","sample"}

\item{cnv_file}{a CNV dataframe with "chromosome","start","end","total_cn","sample"}

\item{cgr}{"connected" CGR regions, output from starfish_link}

\item{genome_v}{genome version of "hg19" or "hg38", default is "hg19"}
}
\value{
plots of "connected" CGR regions.
}
\description{
This function loads a SV dataframe, a CNV dataframe and "connected" CGR regions reported by starfish_link, to draw the connected regions in a linear setting.
}
