% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\docType{data}
\name{starfish_seven_diverging}
\alias{starfish_seven_diverging}
\title{Diverging palette}
\format{
An object of class \code{character} of length 7.
}
\usage{
starfish_seven_diverging
}
\description{
Diverging palette
}
\keyword{datasets}
