% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\docType{data}
\name{ten_colors}
\alias{ten_colors}
\title{Ten-color palette}
\format{
An object of class \code{character} of length 10.
}
\usage{
ten_colors
}
\description{
Ten-color palette
}
\keyword{datasets}
