% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_starfish.R
\name{theme_starfish}
\alias{theme_starfish}
\title{STARFISH ggplot theme}
\usage{
theme_starfish(base_theme = ggplot2::theme_minimal(), ...)
}
\arguments{
\item{base_theme}{A base theme upon which additional theme-specific options are applied}

\item{...}{Fruther arguments passed to \code{ggplot2::theme()}}
}
\description{
\link{ggplot2} plot theme based on...
}
\examples{
\dontrun{
library(ggplot2)
data <- data.frame(x = 1:10, y = 1:10)
ggplot(data, aes(x, y)) + geom_point() + theme_starfish()
}


}
\references{
https://github.com/gadenbuie/ggpomological/blob/master/R/theme_pomological.R
}
\seealso{
\link[ggplot2:theme]{ggplot2::theme}
}
