% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals-ins.R
\docType{data}
\name{wd_data_rd}
\alias{wd_data_rd}
\title{Internal directories}
\format{
An object of class \code{character} of length 1.
}
\usage{
wd_data_rd
}
\description{
paths to the project directory - data folder
}
\keyword{datasets}
