% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals-ins.R
\docType{data}
\name{wd_rd}
\alias{wd_rd}
\title{Internal directories}
\format{
An object of class \code{character} of length 1.
}
\usage{
wd_rd
}
\description{
paths to the project directory
}
\keyword{datasets}
