% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra-palette.R
\name{scale_custom}
\alias{scale_custom}
\alias{scale_color_starfish_d}
\alias{scale_colour_starfish_d}
\alias{scale_fill_starfish_d}
\alias{scale_color_starfish_op}
\alias{scale_colour_starfish_op}
\alias{scale_fill_starfish_op}
\alias{scale_color_starfish_div}
\alias{scale_colour_starfish_div}
\alias{scale_fill_starfish_div}
\alias{scale_color_starfish_c}
\alias{scale_colour_starfish_c}
\alias{scale_fill_starfish_c}
\title{Custom color and fill scales}
\usage{
scale_color_starfish_d(palette = "starfish", extend = FALSE, ...)

scale_colour_starfish_d(palette = "starfish", extend = FALSE, ...)

scale_fill_starfish_d(palette = "starfish", extend = FALSE, ...)

scale_color_starfish_op(
  palette = "starfish",
  val_names = c("good", "neutral", "bad"),
  ...
)

scale_colour_starfish_op(
  palette = "starfish",
  val_names = c("good", "neutral", "bad"),
  ...
)

scale_fill_starfish_op(
  palette = "starfish",
  val_names = c("good", "neutral", "bad"),
  ...
)

scale_color_starfish_div(palette = "starfish", ...)

scale_colour_starfish_div(palette = "starfish", ...)

scale_fill_starfish_div(palette = "starfish", ...)

scale_color_starfish_c(palette = "starfish", ...)

scale_colour_starfish_c(palette = "starfish", ...)

scale_fill_starfish_c(palette = "starfish", ...)
}
\arguments{
\item{palette}{Name of color palette}

\item{extend}{Whether to extend discrete color palette to make sufficient colors for levels needed}

\item{...}{Additional arguments to be passed to internal scale function}

\item{val_names}{For opinionated scales, defaults to "good", "neutral", "bad"}
}
\description{
Custom coloring and filling functions based on unique color palettes
}
\details{
Specific functions include:
\itemize{
\item{scale_(color/colour/fill)_starfish_d}{Discrete palette with either fixed or dynamically extended number of shades}

\item{scale_(color/colour/fill)_starfish_op}{Discrete palette with specific values for "good", "bad", and "neutral"}

\item{scale_(color/colour/fill)_starfish_div}{Continuous diverging color palette, must contain negative, neutral, positive values}

\item{scale_(color/colour/fill)_starfish_c}{Continuous color palette}
}
}
\examples{
library(ggplot2)
ggplot(diamonds[1:2000,], aes(x = cut, y = carat,
color = cut)) +
geom_point() +
scale_color_starfish_d() +
}
